Work around various breaking changes in MSVC and newer versions of CUDA. Update the build target for MSVC 2017. Update the build CUDA build customization to 9.2. Centralize the Maya devkit path to the additional paths instead of putting them in each library path.
This isn't tested to run, but it fixes the build.

+++ src/maya/implicit_surface.cpp
@@ -284,7 +284,7 @@ void ImplicitSurface::load_sampleset(MDataBlock &dataBlock)
         bone->set_enabled(true);
         bone->discard_precompute();
         bone->get_hrbf().init_coeffs(inputSample.nodes, inputSample.n_nodes);
-        printf("update_bone_samples: Solved %i nodes\n", inputSample.nodes.size());
+        printf("update_bone_samples: Solved %i nodes\n", (int) inputSample.nodes.size());
 
         // Make sure the current transforms are applied now that we've changed the bone.
         // XXX: If this is needed, Bone should probably do this internally.
